Fixed product relations, Implemnted image service to save delete image,Completed Product CRUD operation.

+++ src/main/java/com/storelink/repository/ProductRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.storelink.model.Product;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer>,JpaSpecificationExecutor<Product>{
+public interface ProductRepository extends JpaRepository<Product, Long>,JpaSpecificationExecutor<Product>{
 
     
 }
